+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 _Please, add the latest build info on top of the list; use Version::MAJOR only after gold release; keep semantic versioning in line with framework's_
 
+- **0.5.20** [Steve A.] re-sync w/ base engine & API; bundle update
+- **0.5.12** [Steve A.] re-sync w/ base engine
+- **0.5.05** [Steve A.] check team & swimmer presence in meeting before enabling links to results in meetings/_navs; re-sync w/ base engine
 - **0.5.03** [Steve A.] resync w/ base engine: slight change in ImportQueueDecorator so that #chrono_delta_label can be called also on master chrono rows; more spec stability fixes
 - **0.5.02** [Steve A.] added meeting/team_results & meeting/swimmer_results; plus additional specs & feature files & various stability fixes
 - **0.5.01** [Steve A.] full support for reported issues for signed-in users; fixed microtransactions data-import for laps